% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{convertFrameId}
\alias{convertFrameId}
\title{Converts the frame id}
\usage{
convertFrameId(vsc = NULL, R = NULL)
}
\arguments{
\item{vsc}{The frame id as used by vsc}

\item{R}{The frame id as used by R}
}
\value{
The frame id as used by the other program
}
\description{
Converts the frame id form R to vsc or viceversa
}
