% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customVarInfo.R
\name{.vsc.applyVarInfos}
\alias{.vsc.applyVarInfos}
\title{Get info about a variable}
\usage{
.vsc.applyVarInfos(
  v,
  infos = character(0),
  stackingInfos = character(0),
  verbose = getOption("vsc.verboseVarInfos", FALSE),
  ind = NULL
)
}
\arguments{
\item{v}{The variable, any R value.}

\item{infos}{Character vector of field names in \code{varInfo}. Retrieve first match.}

\item{stackingInfos}{Character vector of field names in \code{varInfo}. Retrieve all matches.}

\item{verbose}{Whether to print debug info using \code{logCat}, \code{logPrint}.}

\item{ind}{The indices to retrieve if child variables are retrieved.}
}
\value{
A named list, containing the corresponding \code{varInfo} results.
}
\description{
Get info about a variable by applying the relevant \code{varInfo} entries.
}
\details{
The allowed \code{varInfo} entries and their types are:
\itemize{
\item childVars: MinimalVariable[]
\item nChildVars: number
\item customAttributes: MinmalVariable[]
\item internalAttributes: MinmalVariable[]
\item toString: string
\item type: string
\item evaluateName: string
\item printFunc: function | boolean
}

Where \code{MinimalVariable[]} refers to a list of named lists with entries:
\itemize{
\item name: string
\item rValue: any R value
\item setter: (optional, undocumented)
\item setInfo: (optional, undocumented)
}
}
\seealso{
\code{\link{varInfos}}
}
