% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.cat}
\alias{.vsc.cat}
\title{Modified version of \code{cat()} for vsc}
\usage{
.vsc.cat(...)
}
\arguments{
\item{...}{Arguments passed to base::cat()}
}
\value{
NULL (invisible)
}
\description{
Captures the output of \code{cat(...)} and sends it to vsc together with information about the sourcefile and line
}
