% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.evalInFrame}
\alias{.vsc.evalInFrame}
\title{Evaluate an expression and send result to vsc}
\usage{
.vsc.evalInFrame(expr, frameId, silent = TRUE, id = 0)
}
\arguments{
\item{expr}{The espression to be evaluated}

\item{frameId}{The Id of the frame (as given by vsc)}

\item{id}{The Id of the message sent to vsc}
}
\description{
Evaluates an expression in a given frameId and sends the result to vsc
}
