% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.vsc.getCallingLineAndFile}
\alias{.vsc.getCallingLineAndFile}
\title{Get Line and File info for a frameId}
\usage{
.vsc.getCallingLineAndFile(frameId = 0, skipCalls = 0)
}
\arguments{
\item{frameId}{The id of the frame as used by \code{sys.call()}}

\item{skipCalls}{Skip this number of frames (is substracted from \code{frameId})}

\item{default}{The results defaults to this if an error is encountered}
}
\value{
See \code{.vsc.getLineAndFile}
}
\description{
Get Line and File info for a frameId
}
