% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{.vsc.getDummyStack}
\alias{.vsc.getDummyStack}
\title{Build a dummy stack}
\usage{
.vsc.getDummyStack(dummyFile = NULL)
}
\arguments{
\item{dummyFile}{The file that is returned as the source file}
}
\value{
A stack with the same structure as \code{.vsc.buildStack}
}
\description{
Build a dummy stack that contains only one frame with the .GlobalEnv as only scope.
}
