% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.getFileName}
\alias{.vsc.getFileName}
\title{Get Filename corresponding to a call in a given frame}
\usage{
.vsc.getFileName(call, frame)
}
\arguments{
\item{call}{A function call (usually from the stack)}

\item{frame}{The corresponding frame}
}
\value{
The filename
}
\description{
Get Filename corresponding to a call in a given frame
}
