% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.getLineAndFile}
\alias{.vsc.getLineAndFile}
\title{Get Line and File info for a call}
\usage{
.vsc.getLineAndFile(call, default = NULL)
}
\arguments{
\item{call}{The call for which to find line and file info}

\item{default}{The results defaults to this if an error is encountered}
}
\description{
Get Line and File info for a call
}
