% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.getLineNumber}
\alias{.vsc.getLineNumber}
\title{Get Line-number corresponding to a call in a given frame}
\usage{
.vsc.getLineNumber(call)
}
\arguments{
\item{call}{A function call (usually from the stack)}

\item{frame}{The corresponding frame}
}
\value{
The line-numer
}
\description{
Get Line-number corresponding to a call in a given frame
}
