% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{.vsc.getStack}
\alias{.vsc.getStack}
\title{Build current stack and send to vsc}
\usage{
.vsc.getStack(
  topFrame = parent.frame(),
  id = 0,
  isError = 0,
  dummyFile = NULL,
  forceDummyStack = FALSE,
  includePackages = FALSE
)
}
\arguments{
\item{topFrame}{The first stack frame to consider (= the current function call)}

\item{id}{The id of the message sent to vsc}
}
\value{
None (The current stack, formatted as a nested named list is sent to vsc)
}
\description{
Gives info about the current stack and sends it to vsc
}
