% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.vsc.getVarLists}
\alias{.vsc.getVarLists}
\title{Send info about some vars to vsc}
\usage{
.vsc.getVarLists(refs, id = 0)
}
\arguments{
\item{refs}{A list of variableReferences, as specified in the scopes/previous variables}

\item{id}{The id of the message sent to vsc}
}
\value{
None (The variable info, formatted as a nested named list is sent to vsc)
}
\description{
Gathers info about the specified variablesReferences and sends them to vsc
}
