% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{.vsc.isEvaluating}
\alias{.vsc.isEvaluating}
\title{Check if debugger is evaluating}
\usage{
.vsc.isEvaluating()
}
\value{
Boolean indicating whether an expression is being evaluated
}
\description{
Returns \code{TRUE} iff an expression is being evaluated by the debugger during a breakpoint
}
