% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.lookForMain}
\alias{.vsc.lookForMain}
\title{Look for a main function}
\usage{
.vsc.lookForMain(mainFunction = "main")
}
\arguments{
\item{mainFunction}{Name of the main function, usually "\code{main}"}
}
\description{
Looks for main function and reports back to vsc
}
