% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.vsc.makeStringForVsc}
\alias{.vsc.makeStringForVsc}
\title{Prepare a message as string for vsc}
\usage{
.vsc.makeStringForVsc(message, body = "", id = 0)
}
\arguments{
\item{message}{A string identifying the type of message}

\item{body}{The body of the message. Must be convertible to JSON. Usually named lists.}

\item{id}{The message id. Is usually provided in the function call from vsc.}
}
\value{
A (json) string that can be interpreted by vsc
}
\description{
Prepare a message as string for vsc
}
