% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrites.R
\name{.vsc.message}
\alias{.vsc.message}
\title{Modified version of \code{base::message(...)} for vsc}
\usage{
.vsc.message(
  ...,
  domain = NULL,
  appendLF = TRUE,
  showSource = TRUE,
  skipCalls = 0
)
}
\arguments{
\item{...}{Same as \code{base::message}}

\item{domain}{Same as \code{base::message}}

\item{appendLF}{Same as \code{base::message}}

\item{showSource}{Whether to show the calling source file and line.}

\item{skipCalls}{The number of calls to skip when reporting the calling file and line. Can be used e.g. inside log functions.}
}
\description{
Same as \code{base::message()} but uses \code{.vsc.cat} instead of \code{base::cat}
}
