% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{.vsc.onError}
\alias{.vsc.onError}
\title{Error handler}
\usage{
.vsc.onError(err = NULL)
}
\arguments{
\item{err}{The message to be sent to vsc. Defaults to \code{geterrmessage()}}
}
\description{
Error handler used by vsc. Set with \code{options(error = .vsc.onError)}
}
