% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.prepGlobalEnv}
\alias{.vsc.prepGlobalEnv}
\title{Prep the r session for vsc}
\usage{
.vsc.prepGlobalEnv(
  overwritePrint = TRUE,
  overwriteCat = TRUE,
  overwriteSource = TRUE,
  findMain = TRUE,
  mainFunction = "main",
  allowGlobalDebugging = FALSE,
  rStrings = NULL,
  id = 0
)
}
\arguments{
\item{overwriteSource}{Whether to overwrite \code{base::source} with a version that supports breakpoints}

\item{findMain}{Whether to look for a main function}

\item{mainFunction}{Name of the main function}

\item{allowGlobalDebugging}{Whether to allow debugging after finishing a main function/script}

\item{rStrings}{List of unique strings used to communicate with vsc}

\item{id}{id of the response sent to vsc}

\item{overWritePrint}{Whether to overwrite \code{base::print} with a version that sends output to vsc}

\item{overWriteCat}{Whether to overwrite \code{base::cat} with a version that sends output to vsc}
}
\description{
Preps the r session for runing the r debugger in vscode and reports back to vsc
}
