% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.print}
\alias{.vsc.print}
\title{Modified version of \code{print()} for vsc}
\usage{
.vsc.print(x, ...)
}
\arguments{
\item{...}{Arguments passed to \code{base::cat()}}
}
\value{
\code{NULL} (invisible)
}
\description{
Captures the output of \code{print(...)} and sends it to vsc together with information about the sourcefile and line
}
