% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrites.R
\name{.vsc.refreshBreakpoints}
\alias{.vsc.refreshBreakpoints}
\title{Refresh Breakpoints}
\usage{
.vsc.refreshBreakpoints(envs = NULL)
}
\description{
Refresh breakpoints known to the debugger
Can be used if breakpoints were invalidated by e.g. \code{load_all()} or \code{source()}
}
