% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.sendToVsc}
\alias{.vsc.sendToVsc}
\title{Send a message to vsc}
\usage{
.vsc.sendToVsc(message, body = "", id = 0)
}
\arguments{
\item{message}{A string identifying the type of message}

\item{body}{The body of the message. Must be convertible to JSON. Usually named lists.}

\item{id}{The message id. Is usually provided in the function call from vsc.}
}
\description{
Sends a message (text) together with body and id to vsc
}
