% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{.vsc.setBreakpoints}
\alias{.vsc.setBreakpoints}
\title{Set breakpoints in a file}
\usage{
.vsc.setBreakpoints(file, bps = NULL, includePackages = NULL, id = 0)
}
\arguments{
\item{includePackages}{Whether to set breakpoints in packages}

\item{id}{The id of the answer sent to vsc}

\item{srcfile}{The file in which to set breakpoints}

\item{lines}{A list of lines in which to set breakpoints}

\item{ids}{A list of numbers, specifying the id of each breakpoint. Same length as \code{lines}}
}
\description{
Set breakpoints in a file, and sends a confirmation message to vsc
}
