% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrites.R
\name{.vsc.str}
\alias{.vsc.str}
\title{Modified version of \code{utils::str(...)} for vsc}
\usage{
.vsc.str(object, ..., skipCalls = 0, showSource = TRUE)
}
\arguments{
\item{object}{Same as \code{utils::str}}

\item{...}{Same as \code{utils::str}}

\item{skipCalls}{The number of calls to skip when reporting the calling file and line. Can be used e.g. inside log functions.}

\item{showSource}{Whether to show the calling source file and line.}
}
\description{
Same as \code{utils::str} but uses VS Codes structured variable output.
}
