% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{getDotVars}
\alias{getDotVars}
\title{Get unevaluated variables passed as ellipses}
\usage{
getDotVars(env)
}
\arguments{
\item{env}{The environment where the "..." variable is present}
}
\value{
A named list for each element of "..." containing the
expression that will be evaluated,
the environment where it will be evaluated, and the string
representation of the expression.
}
\description{
Get variables which are passed as ellipses without evaluating them.
}
