% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{getLine}
\alias{getLine}
\title{Get Source Info about a call in a given environment}
\usage{
getLine(frameId)
}
\description{
Get Source Info about a call in a given environment
Get the source line of the function call corresponding to a frameId
}
\details{
Get the source line of the function call corresponding to a frameId
}
