% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackTreeHelpers.R
\name{getLocationInfo}
\alias{getLocationInfo}
\title{Get location info from srcref attribute}
\usage{
getLocationInfo(srcref)
}
\value{
NULL or a LocationInfo object, matching entries for LocationsResponse.body and OutputEvent.body
}
\description{
Get location info from srcref attribute
}
\keyword{internal}
