% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackTreeHelpers.R
\name{getLocationInfoForCall}
\alias{getLocationInfoForCall}
\title{Gather info about a frame's source code}
\usage{
getLocationInfoForCall(call, frameIdR = 0)
}
\arguments{
\item{call}{call object as returned by \code{sys.call}}

\item{frameIdR}{A frame id (as used by R) that can be passed to \code{sys.call}}
}
\value{
NULL or a LocationInfo object, matching entries for LocationsResponse.body and OutputEvent.body
}
\description{
Gathers and returns a named list containing info about the name, path, and content of the source file of a frame
}
\keyword{internal}
