% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStackHelpers.R
\name{getNFrames}
\alias{getNFrames}
\title{Get the number of frames}
\usage{
getNFrames(topFrame)
}
\arguments{
\item{topFrame}{Consider only frames below this frame}
}
\description{
Get the number of frames
}
