% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultVarInfoHelpers.R
\name{getPromiseInfo}
\alias{getPromiseInfo}
\title{Get information about a promise}
\usage{
getPromiseInfo(name, env)
}
\arguments{
\item{name}{The name of the variable}

\item{env}{The environment in which to evaluate}
}
\value{
A named list:
\itemize{
\item \code{code}: the expression that will be evaluated
\item \code{environment}: the environment where the promise is evaluated
}
}
\description{
Get information about a promise
}
\keyword{internal}
