% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackTreeHelpers.R
\name{getScopeEnvs}
\alias{getScopeEnvs}
\title{Get the scopes corresponding to a frame}
\usage{
getScopeEnvs(firstenv = parent.frame(), lastenv = .GlobalEnv)
}
\arguments{
\item{firstenv}{The top environment of the current frame}

\item{lastenv}{The last environment to be considered. By default .GlobalEnv, use emptyenv() to consider package environments}
}
\description{
Gets the scopes corresponding to a frame
}
\keyword{internal}
