% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{getSource}
\alias{getSource}
\title{Gather info about a frame's source code}
\usage{
getSource(frameId)
}
\arguments{
\item{frameId}{A frame id (as used by R) that can be passed to sys.call}
}
\value{
A named list containing info about the source file
}
\description{
Gathers and returns a named list containing info about the name, path, and content of the source file of a frame
}
