% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{getVarListsEntry}
\alias{getVarListsEntry}
\title{Get the variable corresponding to a variablesReference}
\usage{
getVarListsEntry(varRef)
}
\arguments{
\item{varRef}{The variablesReference as returned by \verb{getVarRefForVarListArgs\} or \code\{getVarRefForVar}}
}
\value{
A list of variables corresponding to the varRef
}
\description{
Is basically is a wrapper for \code{session$varLists[[varRef]]}.
If necessary evaluates the call for a given varRef and returns the result.
}
