% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{getVarRefForVarListArgs}
\alias{getVarRefForVarListArgs}
\title{Get VariablesReference for varListArgs}
\usage{
getVarRefForVarListArgs(
  varListArgs = NULL,
  evalCall = FALSE,
  varRef = NULL,
  persistent = FALSE
)
}
\arguments{
\item{varListArgs}{A named list containing the arguments that are passed to \code{getVarList}}

\item{evalCall}{A boolean indicating whether to evaluate the call already (defaults to \code{FALSE})}

\item{varRef}{If specified, this varRef is used instead of incrementing the last varRef by 1}
}
\description{
Gets a variablesReference for a list of arguments meant for \code{getVarList}.
Unless specified the actual call to \code{getVarList} is not evaluated.
Instead the arguments are stored and evaluated when the variablesReference is requested.
}
