% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultVarInfoHelpers.R
\name{isPromise}
\alias{isPromise}
\title{Test if an object is a promise}
\usage{
isPromise(name, env)
}
\arguments{
\item{name}{\code{character(1L)} the name of the object}

\item{env}{\code{\link{environment}} the environment of the object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Only strict (non-forced) promises are considered.
}
\keyword{internal}
