% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communications.R
\name{makeStringForVsc}
\alias{makeStringForVsc}
\title{Prepare a message as string for vsc}
\usage{
makeStringForVsc(body = "")
}
\arguments{
\item{body}{The body of the message. Must be convertible to JSON. Usually named lists.}
}
\value{
A (json) string that can be interpreted by vsc
}
\description{
Prepare a message as string for vsc. DEPRECATED!
}
