% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{makeVarLists}
\alias{makeVarLists}
\title{Get varLists for a list of given varRefs}
\usage{
makeVarLists(refs)
}
\arguments{
\item{refs}{List of varRefs}
}
\description{
Get varLists for a list of given varRefs
}
