% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{NO_getDummyFrame}
\alias{NO_getDummyFrame}
\title{Get a dummy frame}
\usage{
NO_getDummyFrame(frameIdR, frameIdVsc, dummyFile = NULL, lastenv = .GlobalEnv)
}
\arguments{
\item{frameIdR}{The internal id of the frame}

\item{frameIdVsc}{The id of the frame that is reported to vsc}

\item{dummyFile}{The name of the file used as source file}
}
\description{
Returns a dummy frame used in \code{.vsc.getDummyStack}.
}
