% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{NO_getScope}
\alias{NO_getScope}
\title{Gather info about a scope}
\usage{
NO_getScope(env)
}
\arguments{
\item{env}{The environment (=scope) that is to be analyzed}
}
\value{
A named list containing the name of the scope an a variablesReference
}
\description{
Gathers and returns info about the variables in a scope
}
