% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{NO_getScopes}
\alias{NO_getScopes}
\title{Gather info about scopes in a frame}
\usage{
NO_getScopes(frame, lastenv = .GlobalEnv)
}
\arguments{
\item{frame}{A frame as constructed by \code{getStackFrame}}
}
\value{
A list of scopes as constructed by \code{getScope}
}
\description{
Gathers and returns info about the scopes in a frame
}
