% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{NO_getStackFrame}
\alias{NO_getStackFrame}
\title{Gather info about a stack frame}
\usage{
NO_getStackFrame(frameIdR, frameIdVsc)
}
\arguments{
\item{frameIdR}{Frame Id as used by R. Used to identify the frame}

\item{frameIdVsc}{Frame Id as used by vsc. Is only added as info}
}
\description{
Gathers info about a stack frame identified by frameIdR
}
