% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{NO_makeVarLists}
\alias{NO_makeVarLists}
\title{Get varLists for a list of given varRefs}
\usage{
NO_makeVarLists(refs)
}
\arguments{
\item{refs}{List of varRefs}
}
\description{
Get varLists for a list of given varRefs
}
