% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrites.R
\name{outputOverwrites}
\alias{outputOverwrites}
\alias{.vsc.cat}
\alias{.vsc.print}
\alias{.vsc.message}
\alias{.vsc.str}
\title{Modified Output Functions}
\usage{
.vsc.cat(..., skipCalls = 0, showSource = TRUE)

.vsc.print(x, ..., skipCalls = 0, showSource = TRUE)

.vsc.message(
  ...,
  domain = NULL,
  appendLF = TRUE,
  showSource = TRUE,
  skipCalls = 0
)

.vsc.str(object, ..., skipCalls = 0, showSource = TRUE)
}
\arguments{
\item{...}{Further arguments that are passed to the original function}

\item{skipCalls}{The number of calls to skip when reporting the calling file and line. Can be useful e.g. inside a log function.}

\item{showSource}{Whether to show the calling source file and line}

\item{x}{Same as in \code{\link[base:print]{base::print()}}}

\item{domain}{Same as in \code{\link[base:message]{base::message()}}}

\item{appendLF}{Same as in \code{\link[base:message]{base::message()}}}

\item{object}{Same as in \code{\link[utils:str]{utils::str()}}}
}
\value{
These functions return the same value as the overwritten functions themselves.
}
\description{
Modified versions of core R functions that add debugger specific functionality.
With the default debug configuration, the normal R functions are overwritten with these
(using \code{\link[=attach]{attach()}}).

The modified functions are designed to be mostly interchangeable with their counterparts.
If the overwrites are not required, they can be toggled using the launch config entries
\code{overwriteCat}, \code{overwritePrint}, \code{overwriteStr}, and \code{overwriteMessage}.
}
\seealso{
\code{\link[base:cat]{base::cat()}}

\code{\link[base:print]{base::print()}}

\code{\link[base:message]{base::message()}}

\code{\link[utils:str]{utils::str()}}

Other overwrites:
\code{\link[=.vsc.debugSource]{.vsc.debugSource()}},
\code{\link[=.vsc.load_all]{.vsc.load_all()}}
}
\concept{overwrites}
