% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrites.R
\name{printToVsc}
\alias{printToVsc}
\title{Internal function to print to vsc}
\usage{
printToVsc(ret, skipCalls = 0, category = "stdout", showSource = TRUE)
}
\arguments{
\item{ret}{The text to be sent}

\item{skipCalls}{The number of calls to skip when reporting the calling file and line. Can be used e.g. inside log functions.}

\item{category}{The output category ("stdout", "stderr", ...)}

\item{showSource}{Whether to show the calling source file and line.}
}
\description{
Sends text to vsc, together with source information.
Is used by e.g. \link{.vsc.print}, \link{.vsc.cat}
}
\keyword{internal}
