% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communications.R
\name{sendToVsc}
\alias{sendToVsc}
\title{Sends a json to vsc}
\usage{
sendToVsc(body = "")
}
\arguments{
\item{body}{The body of the message. Must be convertible to JSON. Usually named lists.}
}
\description{
Sends a json to vsc
}
