% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq2}
\alias{seq2}
\title{Modified version of base::seq}
\usage{
seq2(from, to=NULL, by=1)
}
\arguments{
\item{from}{Can be the starting value of the sequence, or the end value of the sequence, or a vector of length>1, or a list}

\item{to=NULL}{The ending value of the sequence}

\item{by=1}{The step size (as in \code{base::seq})}
}
\value{
A vector containing a sequence of numbers
}
\description{
Modified version of \code{base::seq}
}
\details{
Basically the same as \code{base::seq}, but returns an empty vector if \code{(to-from)*by<0}.
}
