% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customVarInfo.R
\name{varInfos}
\alias{varInfos}
\alias{.vsc.getAllVarInfoFields}
\alias{.vsc.applyAllVarInfos}
\alias{.vsc.resetVarInfos}
\alias{.vsc.addVarInfo}
\alias{.vsc.removeVarInfos}
\alias{.vsc.getVarInfos}
\alias{.vsc.checkVarInfos}
\title{Tools to modify/debug varInfos}
\usage{
.vsc.getAllVarInfoFields(varInfos = session$varInfos)

.vsc.applyAllVarInfos(v, verbose = TRUE)

.vsc.resetVarInfos()

.vsc.addVarInfo(varInfo, overwrite = FALSE)

.vsc.removeVarInfos(positions)

.vsc.getVarInfos(positions = NULL)

.vsc.checkVarInfos(varInfos = session$varInfos)
}
\arguments{
\item{varInfos}{List of varInfos}

\item{v}{Variable, any R object}

\item{verbose}{passed to \code{.vsc.applyVarInfos}}

\item{varInfo}{A varInfo, i.e. named list.}

\item{overwrite}{Boolean, whether to overwrite the entry in that position}

\item{positions}{Numeric or character vector, the entries to retrieve/remove}
}
\description{
Tools to check the varInfos computed for variables and modify
the list of varInfos used.
}
\seealso{
\code{\link{.vsc.applyVarInfos}}
}
