% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeroList}
\alias{zeroList}
\title{Returns a list of 0s}
\usage{
zeroList(list0)
}
\arguments{
\item{list0}{Any oobject that can be passed to lapply as first argument}
}
\value{
A list of the same length as list0, filled with 0s
}
\description{
Returns a list of 0s, of the same length as the input list
}
