% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communications.R
\name{.vsc.cat}
\alias{.vsc.cat}
\title{Modified version of \code{base::cat()} for vsc}
\usage{
.vsc.cat(..., skipCalls = 0)
}
\arguments{
\item{...}{Arguments passed to base::cat()}

\item{skipCalls}{The number of calls to skip when reporting the calling file and line. Can be used e.g. inside log functions.}
}
\value{
NULL (invisible)
}
\description{
Captures the output of \code{base::cat(...)} and sends it to vsc together with information about the sourcefile and line
}
