% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugSource.R
\name{.vsc.debugSource}
\alias{.vsc.debugSource}
\title{Modified version of \code{base::source}}
\usage{
.vsc.debugSource(
  file,
  local = FALSE,
  envir = NULL,
  chdir = FALSE,
  print.eval = NULL,
  encoding = "unknown",
  ...
)
}
\arguments{
\item{file}{String giving the name of the file to be sourced. Connections etc. are not supported!}

\item{local}{Same as in \code{\link[base:source]{base::source()}}, can be overwritten by specifying an environment in \code{envir}.}

\item{envir}{The environment in which to evaluate the sourced code. Overwrites \code{local}, if specified.}

\item{chdir}{Whether to temporarily change the working directory to the location of \code{file}.}

\item{encoding}{The encoding to be used by \code{\link[base:parse]{base::parse()}}.}

\item{...}{Further arguments are ignored but allowed for compatibility with \code{\link[base:source]{base::source()}}.}
}
\description{
Modified version of \code{base::source} that honors breakpoints set by the debugger.
}
\seealso{
Other overwrites: 
\code{\link{.vsc.cat}()},
\code{\link{.vsc.load_all}()}
}
\concept{overwrites}
