% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{.vsc.evalInFrame}
\alias{.vsc.evalInFrame}
\title{Evaluate an expression and send result to vsc}
\usage{
.vsc.evalInFrame(
  expr,
  frameId,
  silent = TRUE,
  id = 0,
  assignToAns = TRUE,
  catchErrors = TRUE,
  deactivateTracing = silent
)
}
\arguments{
\item{expr}{The espression to be evaluated}

\item{frameId}{The Id of the frame (as given by vsc)}

\item{silent}{Whether to ommit output}

\item{id}{Deprecated}

\item{assignToAns}{Whether to assign the result of the evaluation to .GlobalEnv$.ans}

\item{catchErrors}{Whether to catch errors or let them be handled by \code{options(error = ...)}}

\item{deactivateTracing}{Whether to deactivate tracing (=breakpoints) while evaluating}
}
\description{
Evaluates an expression in a given frameId and sends the result to vsc
}
