% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.vsc.prepGlobalEnv}
\alias{.vsc.prepGlobalEnv}
\title{Runs the \code{main()} function}
\usage{
.vsc.prepGlobalEnv(
  overwritePrint = TRUE,
  overwriteCat = TRUE,
  overwriteSource = TRUE,
  findMain = TRUE,
  mainFunction = "main",
  allowGlobalDebugging = FALSE,
  rStrings = NULL,
  id = 0
)
}
\arguments{
\item{overWritePrint}{Whether to overwrite \code{base::print} with a version that sends output to vsc}

\item{overWriteCat}{Whether to overwrite \code{base::cat} with a version that sends output to vsc}
}
\description{
HAS CHANGED! Following info not up to date!
Looks for a function \code{main()} in the global environment and runs it
Runs main() inside this function, hence \code{parent.frame()} and \code{parent.env()} etc. will behave differently
}
