% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.vsc.print}
\alias{.vsc.print}
\title{Modified version of \code{print()} for vsc}
\usage{
.vsc.print(x, ..., skipCalls = 0)
}
\arguments{
\item{...}{Arguments passed to \code{base::cat()}}

\item{skipCalls}{The number of calls to skip when reporting the calling file and line. Can be used e.g. inside log functions.}
}
\value{
\code{NULL} (invisible)
}
\description{
Captures the output of \code{print(...)} and sends it to vsc together with information about the sourcefile and line
}
