% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{getPromiseInfo}
\alias{getPromiseInfo}
\title{Get information about a promise}
\usage{
getPromiseInfo(name, env)
}
\arguments{
\item{name}{The name of the variable}

\item{env}{The environment in which to evaluate}
}
\value{
A named list containing the expression that will be evaluated, the status whether the promise has already been evaluated, the value if it has already been evaluated, and the environment in which the unevaluated promise will be evaluated.
}
\description{
Get information about a promise
}
\keyword{internal}
