% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{getPromiseVar}
\alias{getPromiseVar}
\title{Get a representation of a promise}
\usage{
getPromiseVar(name, env)
}
\arguments{
\item{name}{The name of the variable}

\item{env}{The environment in which to evaluate}
}
\value{
A named list containing the expression that will be evaluated, the environment in which it will be evaluated and a string representation of the expression
}
\description{
Gets a representation of a promise without forcing the promise
}
