% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultVarInfoHelpers.R
\name{isPromise}
\alias{isPromise}
\title{Test if an object is a promise}
\usage{
isPromise(name, env, strict = TRUE)
}
\arguments{
\item{name}{\code{character(1L)} the name of the object}

\item{env}{\code{\link{environment}} the environment of the object}

\item{strict}{\code{logical(1L)} if \code{strict} is \code{TRUE}
(the default), evaluated promises return with \code{FALSE}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Test if an object is a promise
}
\keyword{internal}
