% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{isPromise}
\alias{isPromise}
\title{Test if in object is a promise}
\usage{
isPromise(name, env, strict = TRUE)
}
\arguments{
\item{name}{\link{character(1L)} the name of the object}

\item{env}{\link{environment} the environment of the object}

\item{strict}{\link{logical(1L)} if \code{strict} is \code{TRUE}
(the default), evaluated promises return with \code{FALSE}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Test if in object is a promise
}
\seealso{
\code{\link{getPromiseVar}}
}
\keyword{internal}
