% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq2}
\alias{seq2}
\title{Modified version of base::seq}
\usage{
seq2(from, to, by=1)

seq2(from)
}
\arguments{
\item{from}{Can be the starting value of the sequence, or the end value of the sequence, or a vector of length>1, or a list}

\item{to}{The ending value of the sequence}

\item{by}{The step size (as in \code{base::seq})}
}
\value{
A vector containing a sequence of numbers
}
\description{
Modified version of \code{base::seq}
}
\details{
If \verb{from, to, by\} are supplied, the function returns the same as \code{base::seq`.
If `from, to} are supplied the function returns \code{NULL} if \code{from>to`,
else the same as `base::seq`.
If `from} is a number, the same as \code\{seq2(1,from)} is returned.
}
