% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{package}
\name{vscDebugger-package}
\alias{vscDebugger-package}
\title{vscDebugger}
\description{
Implementation of the \href{https://microsoft.github.io/debug-adapter-protocol/}{Debug Adapter Protocol} for R
}
\details{
Provides support for the vscode extension \href{https://github.com/ManuelHentschel/VSCode-R-Debugger}{vscode-R-Debugger}.
}
